+++ Modules/HomeMaticBidCoS/PhysicalInterfaces/RTLSDR-LAN.cpp
@@ -154,7 +154,7 @@ void RTLSDR_LAN::listen()
 							break;
 						}
 					}
-				} while(receivedBytes == bufferMax);
+				} while(receivedBytes == (unsigned)bufferMax);
 			}
 			catch(BaseLib::SocketTimeOutException& ex)
 			{
