+++ Modules/MAX/MAX.cpp
@@ -148,7 +148,7 @@ std::shared_ptr<MAXDevice> MAX::getDevice(uint32_t address)
 		_devicesMutex.lock();
 		for(std::vector<std::shared_ptr<BaseLib::Systems::LogicalDevice>>::iterator i = _devices.begin(); i != _devices.end(); ++i)
 		{
-			if((*i)->getAddress() == address)
+			if((uint32_t)(*i)->getAddress() == address)
 			{
 				std::shared_ptr<MAXDevice> device(std::dynamic_pointer_cast<MAXDevice>(*i));
 				if(!device) continue;
